requirements.txt
GitPython>=1.7